(GH-373) Convert pdkNewModule command to feature
This commit converts pdkNewModuleCommand to be part of
PdkCommandFeature.
