--- src/commands/pdk/pdkNewModuleCommand.ts
@@ -1,48 +1,0 @@
-'use strict';
-
-import * as vscode from 'vscode';
-import { ILogger } from '../../logging';
-import { reporter } from '../../telemetry/telemetry';
-import * as messages from '../../messages';
-
-export class PDKNewModuleCommand {
-  private logger: ILogger;
-  private terminal: vscode.Terminal;
-
-  constructor(logger: ILogger, terminal: vscode.Terminal) {
-    this.logger = logger;
-    this.terminal = terminal;
-  }
-
-  public run() {
-    let nameOpts: vscode.QuickPickOptions = {
-      placeHolder: "Enter a name for the new Puppet module",
-      matchOnDescription: true,
-      matchOnDetail: true
-    };
-    let dirOpts: vscode.QuickPickOptions = {
-      placeHolder: "Enter a path for the new Puppet module",
-      matchOnDescription: true,
-      matchOnDetail: true
-    };
-
-    vscode.window.showInputBox(nameOpts).then(moduleName => {
-      if (moduleName == undefined) {
-        vscode.window.showWarningMessage('No module name specifed. Exiting.')
-        return;
-      }
-      vscode.window.showInputBox(dirOpts).then(dir => {
-        this.terminal.sendText(`pdk new module --skip-interview ${moduleName} ${dir}`);
-        this.terminal.sendText(`code ${dir}`)
-        this.terminal.show();
-        if (reporter) {
-          reporter.sendTelemetryEvent(messages.PDKCommandStrings.PdkNewModuleCommandId);
-        }
-      })
-    })
-  }
-
-  public dispose(): any {
-    this.terminal.dispose();
-  }
-}
