+++ src/commands/pdkcommands.ts
@@ -2,19 +2,12 @@ import * as vscode from 'vscode';
 import { PDKCommandStrings } from '../messages';
 import { IConnectionManager } from '../connection';
 import { ILogger } from '../logging';
-import { PDKNewModuleCommand } from './pdk/pdkNewModuleCommand';
 import { PDKNewClassCommand } from './pdk/pdkNewClassCommand';
 import { PDKNewTaskCommand } from './pdk/pdkNewTaskCommand';
 import { PDKValidateCommand } from './pdk/pdkValidateCommand';
 import { PDKTestUnitCommand } from './pdk/pdkTestCommand';
 
 export function setupPDKCommands(langID: string, connManager: IConnectionManager, ctx: vscode.ExtensionContext, logger: ILogger, terminal: vscode.Terminal) {
-  let newModuleCommand = new PDKNewModuleCommand(logger, terminal);
-  ctx.subscriptions.push(newModuleCommand);
-  ctx.subscriptions.push(vscode.commands.registerCommand(PDKCommandStrings.PdkNewModuleCommandId, () => {
-    newModuleCommand.run();
-  }));
-
   let newClassCommand = new PDKNewClassCommand(logger, terminal);
   ctx.subscriptions.push(newClassCommand);
   ctx.subscriptions.push(vscode.commands.registerCommand(PDKCommandStrings.PdkNewClassCommandId, () => {
