+++ src/feature/PdkCommandFeature.ts
@@ -13,9 +13,42 @@ export class PdkCommandFeature implements IFeature {
   constructor(context:vscode.ExtensionContext, logger: ILogger, terminal: vscode.Terminal) {
     this.logger = logger;
     this.terminal = terminal;
+
+    context.subscriptions.push(vscode.commands.registerCommand(PDKCommandStrings.PdkNewModuleCommandId, () => {
+      this.pdkNewModule();
+    }));
+    this.logger.debug("Registered " + PDKCommandStrings.PdkNewModuleCommandId + " command");
   }
 
   public dispose(): any {
     this.terminal.dispose();
+  }
+
+  public pdkNewModule() {
+    let nameOpts: vscode.QuickPickOptions = {
+      placeHolder: "Enter a name for the new Puppet module",
+      matchOnDescription: true,
+      matchOnDetail: true
+    };
+    let dirOpts: vscode.QuickPickOptions = {
+      placeHolder: "Enter a path for the new Puppet module",
+      matchOnDescription: true,
+      matchOnDetail: true
+    };
+
+    vscode.window.showInputBox(nameOpts).then(moduleName => {
+      if (moduleName === undefined) {
+        vscode.window.showWarningMessage('No module name specifed. Exiting.');
+        return;
+      }
+      vscode.window.showInputBox(dirOpts).then(dir => {
+        this.terminal.sendText(`pdk new module --skip-interview ${moduleName} ${dir}`);
+        this.terminal.sendText(`code ${dir}`);
+        this.terminal.show();
+        if (reporter) {
+          reporter.sendTelemetryEvent(PDKCommandStrings.PdkNewModuleCommandId);
+        }
+      });
+    });
   }
 }
